refactor: move resetWasShoveled to its own file
